move offline toast to own component
